% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars2}
\alias{starwars2}
\title{Stars Wars data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 4 columns.
}
\usage{
starwars2
}
\description{
A subset of the starwars data set from the dplyr package
}
\keyword{datasets}
